chore:reorganize-stories-folders
