+++ .storybook/main.ts
@@ -2,7 +2,7 @@ import { withoutVitePlugins } from "@storybook/builder-vite";
 import type { StorybookConfig } from "@storybook/react-vite";
 
 const config: StorybookConfig = {
-    stories: ["../src/stories/*.stories.@(js|jsx|mjs|ts|tsx)"],
+    stories: ["../lib/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
     addons: [
         "@storybook/addon-links",
         "@storybook/addon-essentials",
