+++ lib/components/Typography/text/text.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Text } from "../../lib/components/Typography";
+import { Text } from ".";
 
 const meta = {
     title: "Typography/Text",
